third_party/WebKit/LayoutTests/http/tests/inspector/network/network-empty-xhr.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test() {
  function dumpRequest(request, callback) {
    if (!request)
      return callback();
    TestRunner.addResult(request.url());

    function contentLoaded(content) {
      TestRunner.addResult('resource.content: ' + content);
      callback();
    }

    request.requestContent().then(contentLoaded);
  }

  NetworkTestRunner.recordNetwork();
  NetworkTestRunner.makeSimpleXHR('GET', 'resources/empty.html?sync', false, firstXHRLoaded);

  function firstXHRLoaded() {
    NetworkTestRunner.makeSimpleXHR('GET', 'resources/empty.html?async', true, step2);
  }

  function step2() {
    var requests = NetworkTestRunner.networkRequests();
    var request = requests[requests.length - 2];
    dumpRequest(request, step3);
  }

  function step3() {
    var requests = NetworkTestRunner.networkRequests();
    var request = requests[requests.length - 1];
    dumpRequest(request, step4);
  }

  function step4() {
    TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
<p> Tests empty xhr content is correctly loaded in inspector.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=79026">Bug 79026</a>
</body>
</html>
