third_party/WebKit/LayoutTests/http/tests/inspector/network/network-empty-xhr-expected.txt

CONSOLE WARNING: Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
CONSOLE MESSAGE: line 6: XHR loaded: 1
CONSOLE MESSAGE: line 6: XHR loaded: 2
Tests empty xhr content is correctly loaded in inspector.

Bug 79026
http://127.0.0.1:8000/inspector/network/resources/empty.html?sync
resource.content: 
http://127.0.0.1:8000/inspector/network/resources/empty.html?async
resource.content: 
